+++ special-pages/shared/handlers.js
@@ -1,0 +1,14 @@
+/**
+ * @param {MouseEvent} event
+ * @param {ImportMeta['platform']} platformName
+ * @return {'new-tab' | 'new-window' | 'same-tab'}
+ */
+export function eventToTarget(event, platformName) {
+    const isControlClick = platformName === 'macos' ? event.metaKey : event.ctrlKey;
+    if (isControlClick) {
+        return 'new-tab';
+    } else if (event.shiftKey || event.button === 1 /* middle click */) {
+        return 'new-window';
+    }
+    return 'same-tab';
+}